# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery Management API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Order {
  id: string;
  customer_name: string;
  phone_number: string;
  category: string;
  item_name: string;
  quantity: number;
  total_amount: number;
  advance_payment: number;
  remaining_balance: number;
  delivery_date: string;
  order_status: string;
  payment_status: string;
  /** @nullable */
  notes?: string | null;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
}

export interface OrderInput {
  customer_name: string;
  phone_number: string;
  category: string;
  item_name: string;
  quantity: number;
  total_amount: number;
  advance_payment: number;
  delivery_date: string;
  order_status?: string;
  payment_status?: string;
  notes?: string;
}

export interface OrderUpdate {
  customer_name?: string;
  phone_number?: string;
  category?: string;
  item_name?: string;
  quantity?: number;
  total_amount?: number;
  advance_payment?: number;
  delivery_date?: string;
  order_status?: string;
  payment_status?: string;
  notes?: string;
}

export interface Customer {
  id: string;
  name: string;
  phone_number: string;
  total_orders: number;
  total_spent: number;
  /** @nullable */
  last_order_date?: string | null;
}

export interface DashboardStats {
  today_orders: number;
  pending_orders: number;
  completed_orders: number;
  total_revenue: number;
  pending_payments: number;
  total_customers: number;
  this_month_revenue: number;
  delivery_today: number;
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface CategoryBreakdown {
  category: string;
  count: number;
  revenue: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  message: string;
  timestamp: string;
  /** @nullable */
  customer_name?: string | null;
  /** @nullable */
  amount?: number | null;
}

export type ListOrdersParams = {
  status?: string;
  payment_status?: string;
  search?: string;
  date_from?: string;
  date_to?: string;
};

export type ListCustomersParams = {
  search?: string;
};

export type GetRevenueChartParams = {
  days?: number;
};