# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  type: string;
  message: string;
  timestamp: string;
  /** @nullable */
  customer_name?: string | null;
  /** @nullable */
  amount?: number | null;
}