# lib/api-zod/src/generated/types/categoryBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryBreakdown {
  category: string;
  count: number;
  revenue: number;
}