# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: string;
  name: string;
  phone_number: string;
  total_orders: number;
  total_spent: number;
  /** @nullable */
  last_order_date?: string | null;
}