# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  today_orders: number;
  pending_orders: number;
  completed_orders: number;
  total_revenue: number;
  pending_payments: number;
  total_customers: number;
  this_month_revenue: number;
  delivery_today: number;
}