# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  customer_name: string;
  phone_number: string;
  category: string;
  item_name: string;
  quantity: number;
  total_amount: number;
  advance_payment: number;
  delivery_date: string;
  order_status?: string;
  payment_status?: string;
  notes?: string;
}