# lib/api-zod/src/generated/types/revenueDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bakery Management API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RevenueDataPoint {
  date: string;
  revenue: number;
  orders: number;
}